Split optimizer tests into individual files
Unit tests for most optimizer classes were all located in a single test
file `test_optimize.py`, making it unwieldy. This file now only contains
tests common to all optimizers.
