+++ tests/optimize/test_gradient_descent.py
@@ -1,0 +1,241 @@
+# Copyright 2018-2021 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+Unit tests for the ``GradientDescentOptimizer``.
+"""
+import pytest
+
+import pennylane as qml
+from pennylane import numpy as np
+from pennylane.optimize import GradientDescentOptimizer
+
+x_vals = np.linspace(-10, 10, 16, endpoint=False)
+
+# Hyperparameters for optimizers
+stepsize = 0.1
+
+# function arguments in various formats
+mixed_list = [(0.2, 0.3), np.array([0.4, 0.2, 0.4]), 0.1]
+multid_array = np.array([[0.1, 0.2], [-0.1, -0.4]])
+multid_list = [[0.1, 0.2], [-0.1, -0.4]]
+
+# functions and their gradients
+fnames = ["test_function_1", "test_function_2", "test_function_3"]
+univariate_funcs = [np.sin, lambda x: np.exp(x / 10.0), lambda x: x ** 2]
+grad_uni_fns = [lambda x: (np.cos(x),), lambda x: (np.exp(x / 10.0) / 10.0,), lambda x: (2 * x,)]
+
+multivariate_funcs = [
+    lambda x: np.sin(x[0]) + np.cos(x[1]),
+    lambda x: np.exp(x[0] / 3) * np.tanh(x[1]),
+    lambda x: np.sum([x_ ** 2 for x_ in x]),
+]
+grad_multi_funcs = [
+    lambda x: (np.array([np.cos(x[0]), -np.sin(x[1])]),),
+    lambda x: (
+        np.array(
+            [np.exp(x[0] / 3) / 3 * np.tanh(x[1]), np.exp(x[0] / 3) * (1 - np.tanh(x[1]) ** 2)]
+        ),
+    ),
+    lambda x: (np.array([2 * x_ for x_ in x]),),
+]
+
+mvar_mdim_funcs = [
+    lambda x: np.sin(x[0, 0]) + np.cos(x[1, 0]) - np.sin(x[0, 1]) + x[1, 1],
+    lambda x: np.exp(x[0, 0] / 3) * np.tanh(x[0, 1]),
+    lambda x: np.sum([x_[0] ** 2 for x_ in x]),
+]
+grad_mvar_mdim_funcs = [
+    lambda x: (np.array([[np.cos(x[0, 0]), -np.cos(x[0, 1])], [-np.sin(x[1, 0]), 1.0]]),),
+    lambda x: (
+        np.array(
+            [
+                [
+                    np.exp(x[0, 0] / 3) / 3 * np.tanh(x[0, 1]),
+                    np.exp(x[0, 0] / 3) * (1 - np.tanh(x[0, 1]) ** 2),
+                ],
+                [0.0, 0.0],
+            ]
+        ),
+    ),
+    lambda x: (np.array([[2 * x_[0], 0.0] for x_ in x]),),
+]
+
+
+@qml.qnode(qml.device("default.qubit", wires=1))
+def quant_fun(*variables):
+    qml.RX(variables[0][1], wires=[0])
+    qml.RY(variables[1][2], wires=[0])
+    qml.RY(variables[2], wires=[0])
+    return qml.expval(qml.PauliZ(0))
+
+
+@qml.qnode(qml.device("default.qubit", wires=1))
+def quant_fun_mdarr(var):
+    qml.RX(var[0, 1], wires=[0])
+    qml.RY(var[1, 0], wires=[0])
+    qml.RY(var[1, 1], wires=[0])
+    return qml.expval(qml.PauliZ(0))
+
+
+@qml.qnode(qml.device("default.qubit", wires=1))
+def quant_fun_mdlist(var):
+    qml.RX(var[0][1], wires=[0])
+    qml.RY(var[1][0], wires=[0])
+    qml.RY(var[1][1], wires=[0])
+    return qml.expval(qml.PauliZ(0))
+
+
+class TestGradientDescentOptimizer:
+    """Test the Gradient Descent optimizer"""
+
+    def test_array_and_list_return_same_update(self, tol):
+        """Tests that gradient descent optimizer has the same output for
+        lists and arrays."""
+        sgd_opt = GradientDescentOptimizer(stepsize)
+
+        def hybrid_fun_mdarr(var):
+            return quant_fun_mdarr(var) + var[0, 0]
+
+        def hybrid_fun_mdlist(var):
+            return quant_fun_mdlist(var) + var[0][0]
+
+        array = sgd_opt.step(hybrid_fun_mdarr, multid_array)
+        ls = sgd_opt.step(hybrid_fun_mdlist, multid_list)
+
+        assert np.allclose(array, np.asarray(ls), atol=tol)
+
+    @pytest.mark.parametrize("func, f_grad", list(zip(univariate_funcs, grad_uni_fns)))
+    @pytest.mark.parametrize("var", [0, -3, 42])
+    def test_step_and_cost_supplied_grad(self, func, var, f_grad):
+        """Test that returned cost is correct if gradient function is supplied"""
+        sgd_opt = GradientDescentOptimizer(stepsize)
+
+        _, res = sgd_opt.step_and_cost(func, var, grad_fn=f_grad)
+        expected = func(var)
+
+        assert np.all(res == expected)
+
+    def test_step_and_cost_autograd_sgd_mixed_list(self):
+        """Test that the correct cost is returned via the step_and_cost method for the
+        gradient-descent optimizer"""
+        sgd_opt = GradientDescentOptimizer(stepsize)
+
+        _, res = sgd_opt.step_and_cost(quant_fun, *mixed_list)
+        expected = quant_fun(*mixed_list)
+
+        assert np.all(res == expected)
+
+    def test_step_and_cost_autograd_sgd_multid_array(self):
+        """Test that the correct cost is returned via the step_and_cost method for the
+        gradient-descent optimizer"""
+        sgd_opt = GradientDescentOptimizer(stepsize)
+
+        _, res = sgd_opt.step_and_cost(quant_fun_mdarr, multid_array)
+        expected = quant_fun_mdarr(multid_array)
+
+        assert np.all(res == expected)
+
+    @pytest.mark.parametrize("x_start", x_vals)
+    def test_gradient_descent_optimizer_univar(self, x_start, tol):
+        """Tests that basic stochastic gradient descent takes gradient-descent steps correctly
+        for univariate functions."""
+        sgd_opt = GradientDescentOptimizer(stepsize)
+
+        # TODO parametrize this for also
+        for gradf, f, _ in zip(grad_uni_fns, univariate_funcs, fnames):
+            x_new = sgd_opt.step(f, x_start)
+            x_correct = x_start - gradf(x_start)[0] * stepsize
+            assert np.allclose(x_new, x_correct, atol=tol)
+
+    def test_gradient_descent_optimizer_multivar(self, tol):
+        """Tests that basic stochastic gradient descent takes gradient-descent steps correctly
+        for multivariate functions."""
+        sgd_opt = GradientDescentOptimizer(stepsize)
+
+        for gradf, f, _ in zip(grad_multi_funcs, multivariate_funcs, fnames):
+            for jdx in range(len(x_vals[:-1])):
+                x_vec = x_vals[jdx : jdx + 2]
+                x_new = sgd_opt.step(f, x_vec)
+                x_correct = x_vec - gradf(x_vec)[0] * stepsize
+                assert np.allclose(x_new, x_correct, atol=tol)
+
+    def test_gradient_descent_optimizer_multivar_multidim(self, tol):
+        """Tests that basic stochastic gradient descent takes gradient-descent steps correctly
+        for multivariate functions and with higher dimensional inputs."""
+        sgd_opt = GradientDescentOptimizer(stepsize)
+
+        for gradf, f, _ in zip(grad_mvar_mdim_funcs, mvar_mdim_funcs, fnames):
+            for jdx in range(len(x_vals[:-3])):
+                x_vec = x_vals[jdx : jdx + 4]
+                x_vec_multidim = np.reshape(x_vec, (2, 2))
+                x_new = sgd_opt.step(f, x_vec_multidim)
+                x_correct = x_vec_multidim - gradf(x_vec_multidim)[0] * stepsize
+                x_new_flat = x_new.flatten()
+                x_correct_flat = x_correct.flatten()
+                assert np.allclose(x_new_flat, x_correct_flat, atol=tol)
+
+    @pytest.mark.parametrize("x_start", x_vals)
+    def test_gradient_descent_optimizer_usergrad(self, x_start, tol):
+        """Tests that basic stochastic gradient descent takes gradient-descent steps correctly
+        using user-provided gradients."""
+        sgd_opt = GradientDescentOptimizer(stepsize)
+
+        for gradf, f, _ in zip(grad_uni_fns[::-1], univariate_funcs, fnames):
+            x_new = sgd_opt.step(f, x_start, grad_fn=gradf)
+            x_correct = x_start - gradf(x_start)[0] * stepsize
+            assert np.allclose(x_new, x_correct, atol=tol)
+
+    def test_update_stepsize(self):
+        """
+        Tests whether the stepsize value is updated correctly and whether a ``UserWarning``
+        is raised when the ``update_stepsize`` method is used.
+        """
+
+        eta = 0.5
+        opt = GradientDescentOptimizer(eta)
+        assert opt.stepsize == eta
+
+        eta2 = 0.1
+        opt.update_stepsize(eta2)
+        assert opt.stepsize == eta2
+
+        with pytest.warns(
+            UserWarning,
+            match="'update_stepsize' is deprecated. Stepsize value can be updated using "
+            "the 'stepsize' attribute.",
+        ):
+            opt.update_stepsize(eta)
+
+    def test_private_stepsize(self):
+        """
+        Tests whether it is possible to get and set ``stepsize`` using ``_stepsize``
+        and whether a ``UserWarning`` is raised while doing so.
+        """
+        eta = 0.5
+        opt = GradientDescentOptimizer(eta)
+        assert opt._stepsize == eta
+
+        with pytest.warns(
+            UserWarning, match="'_stepsize' is deprecated. Please use 'stepsize' instead."
+        ):
+            opt._stepsize
+
+        eta2 = 0.1
+        opt._stepsize = eta2
+        assert opt.stepsize == eta2
+
+        with pytest.warns(
+            UserWarning, match="'_stepsize' is deprecated. Please use 'stepsize' instead."
+        ):
+            opt._stepsize = eta2
