+++ tests/optimize/test_optimize.py
@@ -15,9 +15,6 @@
 Unit tests for the :mod:`pennylane` optimizers.
 """
 # pylint: disable=redefined-outer-name
-import itertools as it
-
-import numpy as onp
 import pytest
 
 import pennylane as qml
@@ -29,670 +26,13 @@ from pennylane.optimize import (
     AdagradOptimizer,
     RMSPropOptimizer,
     AdamOptimizer,
-    RotoselectOptimizer,
     RotosolveOptimizer,
 )
-
-
-x_vals = np.linspace(-10, 10, 16, endpoint=False)
 
 # Hyperparameters for optimizers
 stepsize = 0.1
 gamma = 0.5
 delta = 0.8
-
-
-# function arguments in various formats
-mixed_list = [(0.2, 0.3), np.array([0.4, 0.2, 0.4]), 0.1]
-flat_list = [0.2, 0.3, 0.1, 0.4, -0.1]
-multid_array = np.array([[0.1, 0.2], [-0.1, -0.4]])
-multid_list = [[0.1, 0.2], [-0.1, -0.4]]
-
-
-# functions and their gradients
-fnames = ["test_function_1", "test_function_2", "test_function_3"]
-univariate_funcs = [np.sin, lambda x: np.exp(x / 10.0), lambda x: x ** 2]
-grad_uni_fns = [lambda x: (np.cos(x),), lambda x: (np.exp(x / 10.0) / 10.0,), lambda x: (2 * x,)]
-
-multivariate_funcs = [
-    lambda x: np.sin(x[0]) + np.cos(x[1]),
-    lambda x: np.exp(x[0] / 3) * np.tanh(x[1]),
-    lambda x: np.sum([x_ ** 2 for x_ in x]),
-]
-grad_multi_funcs = [
-    lambda x: (np.array([np.cos(x[0]), -np.sin(x[1])]),),
-    lambda x: (
-        np.array(
-            [np.exp(x[0] / 3) / 3 * np.tanh(x[1]), np.exp(x[0] / 3) * (1 - np.tanh(x[1]) ** 2)]
-        ),
-    ),
-    lambda x: (np.array([2 * x_ for x_ in x]),),
-]
-
-mvar_mdim_funcs = [
-    lambda x: np.sin(x[0, 0]) + np.cos(x[1, 0]) - np.sin(x[0, 1]) + x[1, 1],
-    lambda x: np.exp(x[0, 0] / 3) * np.tanh(x[0, 1]),
-    lambda x: np.sum([x_[0] ** 2 for x_ in x]),
-]
-grad_mvar_mdim_funcs = [
-    lambda x: (np.array([[np.cos(x[0, 0]), -np.cos(x[0, 1])], [-np.sin(x[1, 0]), 1.0]]),),
-    lambda x: (
-        np.array(
-            [
-                [
-                    np.exp(x[0, 0] / 3) / 3 * np.tanh(x[0, 1]),
-                    np.exp(x[0, 0] / 3) * (1 - np.tanh(x[0, 1]) ** 2),
-                ],
-                [0.0, 0.0],
-            ]
-        ),
-    ),
-    lambda x: (np.array([[2 * x_[0], 0.0] for x_ in x]),),
-]
-
-
-@qml.qnode(qml.device("default.qubit", wires=1))
-def quant_fun(*variables):
-    qml.RX(variables[0][1], wires=[0])
-    qml.RY(variables[1][2], wires=[0])
-    qml.RY(variables[2], wires=[0])
-    return qml.expval(qml.PauliZ(0))
-
-
-@qml.qnode(qml.device("default.qubit", wires=1))
-def quant_fun_flat(var):
-    qml.RX(var[0], wires=[0])
-    qml.RY(var[1], wires=[0])
-    qml.RY(var[2], wires=[0])
-    qml.RX(var[3], wires=[0])
-    return qml.expval(qml.PauliZ(0))
-
-
-@qml.qnode(qml.device("default.qubit", wires=1))
-def quant_fun_mdarr(var):
-    qml.RX(var[0, 1], wires=[0])
-    qml.RY(var[1, 0], wires=[0])
-    qml.RY(var[1, 1], wires=[0])
-    return qml.expval(qml.PauliZ(0))
-
-
-@qml.qnode(qml.device("default.qubit", wires=1))
-def quant_fun_mdlist(var):
-    qml.RX(var[0][1], wires=[0])
-    qml.RY(var[1][0], wires=[0])
-    qml.RY(var[1][1], wires=[0])
-    return qml.expval(qml.PauliZ(0))
-
-
-@pytest.fixture(scope="function")
-def bunch():
-    class A:
-        sgd_opt = GradientDescentOptimizer(stepsize)
-        mom_opt = MomentumOptimizer(stepsize, momentum=gamma)
-        nesmom_opt = NesterovMomentumOptimizer(stepsize, momentum=gamma)
-        adag_opt = AdagradOptimizer(stepsize)
-        rms_opt = RMSPropOptimizer(stepsize, decay=gamma)
-        adam_opt = AdamOptimizer(stepsize, beta1=gamma, beta2=delta)
-        rotoselect_opt = RotoselectOptimizer()
-
-    return A()
-
-
-class TestOptimizer:
-    """Basic optimizer tests."""
-
-    def test_array_and_list_return_same_update(self, bunch, tol):
-        """Tests that gradient descent optimizer has the same output for
-        lists and arrays."""
-
-        def hybrid_fun_mdarr(var):
-            return quant_fun_mdarr(var) + var[0, 0]
-
-        def hybrid_fun_mdlist(var):
-            return quant_fun_mdlist(var) + var[0][0]
-
-        array = bunch.sgd_opt.step(hybrid_fun_mdarr, multid_array)
-        ls = bunch.sgd_opt.step(hybrid_fun_mdlist, multid_list)
-
-        assert np.allclose(array, np.asarray(ls), atol=tol)
-
-    def test_step_and_cost_autograd_sgd_mixed_list(self, bunch):
-        """Test that the correct cost is returned via the step_and_cost method for the
-        gradient-descent optimizer"""
-        _, res = bunch.sgd_opt.step_and_cost(quant_fun, *mixed_list)
-        expected = quant_fun(*mixed_list)
-
-        assert np.all(res == expected)
-
-    def test_step_and_cost_autograd_sgd_multid_array(self, bunch):
-        """Test that the correct cost is returned via the step_and_cost method for the
-        gradient-descent optimizer"""
-        _, res = bunch.sgd_opt.step_and_cost(quant_fun_mdarr, multid_array)
-        expected = quant_fun_mdarr(multid_array)
-
-        assert np.all(res == expected)
-
-    def test_step_and_cost_autograd_nesterov_mixed_list(self, bunch):
-        """Test that the correct cost is returned via the step_and_cost method for the
-        Nesterov momentum optimizer"""
-        _, res = bunch.nesmom_opt.step_and_cost(quant_fun, *mixed_list)
-        expected = quant_fun(*mixed_list)
-
-        assert np.all(res == expected)
-
-    def test_step_and_cost_autograd_nesterov_multid_array(self, bunch):
-        """Test that the correct cost is returned via the step_and_cost method for the
-        Nesterov momentum optimizer"""
-        _, res = bunch.nesmom_opt.step_and_cost(quant_fun_mdarr, multid_array)
-        expected = quant_fun_mdarr(multid_array)
-
-        assert np.all(res == expected)
-
-    @pytest.mark.parametrize("params", [[1.7, 2.2], [-1.42, 0.1], [0.05, -0.8]])
-    def test_step_and_cost_autograd_rotoselect(self, bunch, params):
-        """Test that the correct cost is returned via the step_and_cost method for the
-        Rotoselect momentum optimizer"""
-        generators = [qml.RY, qml.RX]
-        possible_generators = [qml.RX, qml.RY, qml.RZ]
-        bunch.rotoselect_opt.possible_generators = possible_generators
-
-        dev = qml.device("default.qubit", shots=None, wires=2)
-
-        def ansatz(params, generators):
-            generators[0](params[0], wires=0)
-            generators[1](params[1], wires=1)
-            qml.CNOT(wires=[0, 1])
-
-        @qml.qnode(dev)
-        def circuit_1(params, generators=None):  # generators will be passed as a keyword arg
-            ansatz(params, generators)
-            return qml.expval(qml.PauliZ(0)), qml.expval(qml.PauliY(1))
-
-        @qml.qnode(dev)
-        def circuit_2(params, generators=None):  # generators will be passed as a keyword arg
-            ansatz(params, generators)
-            return qml.expval(qml.PauliX(0))
-
-        def cost_fn(params, generators):
-            Z_1, Y_2 = circuit_1(params, generators=generators)
-            X_1 = circuit_2(params, generators=generators)
-            return 0.5 * Y_2 + 0.8 * Z_1 - 0.2 * X_1
-
-        _, _, res = bunch.rotoselect_opt.step_and_cost(cost_fn, params, generators)
-        expected = cost_fn(params, generators)
-
-        assert np.all(res == expected)
-
-    @pytest.mark.parametrize("func, f_grad", list(zip(univariate_funcs, grad_uni_fns)))
-    @pytest.mark.parametrize("var", [0, -3, 42])
-    def test_step_and_cost_supplied_grad(self, bunch, func, var, f_grad):
-        """Test that returned cost is correct if gradient function is supplied"""
-        _, res = bunch.sgd_opt.step_and_cost(func, var, grad_fn=f_grad)
-        expected = func(var)
-
-        assert np.all(res == expected)
-
-    @pytest.mark.parametrize("x_start", x_vals)
-    def test_gradient_descent_optimizer_univar(self, x_start, bunch, tol):
-        """Tests that basic stochastic gradient descent takes gradient-descent steps correctly
-        for univariate functions."""
-
-        # TODO parametrize this for also
-        for gradf, f, name in zip(grad_uni_fns, univariate_funcs, fnames):
-            x_new = bunch.sgd_opt.step(f, x_start)
-            x_correct = x_start - gradf(x_start)[0] * stepsize
-            assert np.allclose(x_new, x_correct, atol=tol)
-
-    def test_gradient_descent_optimizer_multivar(self, bunch, tol):
-        """Tests that basic stochastic gradient descent takes gradient-descent steps correctly
-        for multivariate functions."""
-
-        for gradf, f, name in zip(grad_multi_funcs, multivariate_funcs, fnames):
-            for jdx in range(len(x_vals[:-1])):
-                x_vec = x_vals[jdx : jdx + 2]
-                x_new = bunch.sgd_opt.step(f, x_vec)
-                x_correct = x_vec - gradf(x_vec)[0] * stepsize
-                assert np.allclose(x_new, x_correct, atol=tol)
-
-    def test_gradient_descent_optimizer_multivar_multidim(self, bunch, tol):
-        """Tests that basic stochastic gradient descent takes gradient-descent steps correctly
-        for multivariate functions and with higher dimensional inputs."""
-
-        for gradf, f, name in zip(grad_mvar_mdim_funcs, mvar_mdim_funcs, fnames):
-            for jdx in range(len(x_vals[:-3])):
-                x_vec = x_vals[jdx : jdx + 4]
-                x_vec_multidim = np.reshape(x_vec, (2, 2))
-                x_new = bunch.sgd_opt.step(f, x_vec_multidim)
-                x_correct = x_vec_multidim - gradf(x_vec_multidim)[0] * stepsize
-                x_new_flat = x_new.flatten()
-                x_correct_flat = x_correct.flatten()
-                assert np.allclose(x_new_flat, x_correct_flat, atol=tol)
-
-    @pytest.mark.parametrize("x_start", x_vals)
-    def test_gradient_descent_optimizer_usergrad(self, x_start, bunch, tol):
-        """Tests that basic stochastic gradient descent takes gradient-descent steps correctly
-        using user-provided gradients."""
-
-        for gradf, f, name in zip(grad_uni_fns[::-1], univariate_funcs, fnames):
-            x_new = bunch.sgd_opt.step(f, x_start, grad_fn=gradf)
-            x_correct = x_start - gradf(x_start)[0] * stepsize
-            assert np.allclose(x_new, x_correct, atol=tol)
-
-    @pytest.mark.parametrize("x_start", x_vals)
-    def test_momentum_optimizer_univar(self, x_start, bunch, tol):
-        """Tests that momentum optimizer takes one and two steps correctly
-        for univariate functions."""
-
-        for gradf, f, name in zip(grad_uni_fns, univariate_funcs, fnames):
-            bunch.mom_opt.reset()
-
-            x_onestep = bunch.mom_opt.step(f, x_start)
-            x_onestep_target = x_start - gradf(x_start)[0] * stepsize
-            assert np.allclose(x_onestep, x_onestep_target, atol=tol)
-
-            x_twosteps = bunch.mom_opt.step(f, x_onestep)
-            momentum_term = gamma * gradf(x_start)[0]
-            x_twosteps_target = x_onestep - (gradf(x_onestep)[0] + momentum_term) * stepsize
-            assert np.allclose(x_twosteps, x_twosteps_target, atol=tol)
-
-    def test_momentum_optimizer_multivar(self, bunch, tol):
-        """Tests that momentum optimizer takes one and two steps correctly
-        for multivariate functions."""
-
-        for gradf, f, name in zip(grad_multi_funcs, multivariate_funcs, fnames):
-            for jdx in range(len(x_vals[:-1])):
-                bunch.mom_opt.reset()
-
-                x_vec = x_vals[jdx : jdx + 2]
-                x_onestep = bunch.mom_opt.step(f, x_vec)
-                x_onestep_target = x_vec - gradf(x_vec)[0] * stepsize
-                assert np.allclose(x_onestep, x_onestep_target, atol=tol)
-
-                x_twosteps = bunch.mom_opt.step(f, x_onestep)
-                momentum_term = gamma * gradf(x_vec)[0]
-                x_twosteps_target = x_onestep - (gradf(x_onestep)[0] + momentum_term) * stepsize
-                assert np.allclose(x_twosteps, x_twosteps_target, atol=tol)
-
-    @pytest.mark.parametrize("x_start", x_vals)
-    def test_nesterovmomentum_optimizer_univar(self, x_start, bunch, tol):
-        """Tests that nesterov momentum optimizer takes one and two steps correctly
-        for univariate functions."""
-
-        for gradf, f, name in zip(grad_uni_fns, univariate_funcs, fnames):
-            bunch.nesmom_opt.reset()
-
-            x_onestep = bunch.nesmom_opt.step(f, x_start)
-            x_onestep_target = x_start - gradf(x_start)[0] * stepsize
-            assert np.allclose(x_onestep, x_onestep_target, atol=tol)
-
-            x_twosteps = bunch.nesmom_opt.step(f, x_onestep)
-            momentum_term = gamma * gradf(x_start)[0]
-            shifted_grad_term = gradf(x_onestep - stepsize * momentum_term)[0]
-            x_twosteps_target = x_onestep - (shifted_grad_term + momentum_term) * stepsize
-            assert np.allclose(x_twosteps, x_twosteps_target, atol=tol)
-
-    def test_nesterovmomentum_optimizer_multivar(self, bunch, tol):
-        """Tests that nesterov momentum optimizer takes one and two steps correctly
-        for multivariate functions."""
-
-        for gradf, f, name in zip(grad_multi_funcs, multivariate_funcs, fnames):
-            for jdx in range(len(x_vals[:-1])):
-                bunch.nesmom_opt.reset()
-
-                x_vec = x_vals[jdx : jdx + 2]
-                x_onestep = bunch.nesmom_opt.step(f, x_vec)
-                x_onestep_target = x_vec - gradf(x_vec)[0] * stepsize
-                assert np.allclose(x_onestep, x_onestep_target, atol=tol)
-
-                x_twosteps = bunch.nesmom_opt.step(f, x_onestep)
-                momentum_term = gamma * gradf(x_vec)[0]
-                shifted_grad_term = gradf(x_onestep - stepsize * momentum_term)[0]
-                x_twosteps_target = x_onestep - (shifted_grad_term + momentum_term) * stepsize
-                assert np.allclose(x_twosteps, x_twosteps_target, atol=tol)
-
-    @pytest.mark.parametrize("x_start", x_vals)
-    def test_nesterovmomentum_optimizer_usergrad(self, x_start, bunch, tol):
-        """Tests that nesterov momentum optimizer takes gradient-descent steps correctly
-        using user-provided gradients."""
-
-        for gradf, f, name in zip(grad_uni_fns[::-1], univariate_funcs, fnames):
-            bunch.nesmom_opt.reset()
-
-            x_onestep = bunch.nesmom_opt.step(f, x_start, grad_fn=gradf)
-            x_onestep_target = x_start - gradf(x_start)[0] * stepsize
-            assert np.allclose(x_onestep, x_onestep_target, atol=tol)
-
-            x_twosteps = bunch.nesmom_opt.step(f, x_onestep, grad_fn=gradf)
-            momentum_term = gamma * gradf(x_start)[0]
-            shifted_grad_term = gradf(x_onestep - stepsize * momentum_term)[0]
-            x_twosteps_target = x_onestep - (shifted_grad_term + momentum_term) * stepsize
-            assert np.allclose(x_twosteps, x_twosteps_target, atol=tol)
-
-    @pytest.mark.parametrize("x_start", x_vals)
-    def test_adagrad_optimizer_univar(self, x_start, bunch, tol):
-        """Tests that adagrad optimizer takes one and two steps correctly
-        for univariate functions."""
-
-        for gradf, f, name in zip(grad_uni_fns, univariate_funcs, fnames):
-            bunch.adag_opt.reset()
-
-            x_onestep = bunch.adag_opt.step(f, x_start)
-            past_grads = gradf(x_start)[0] * gradf(x_start)[0]
-            adapt_stepsize = stepsize / np.sqrt(past_grads + 1e-8)
-            x_onestep_target = x_start - gradf(x_start)[0] * adapt_stepsize
-            assert np.allclose(x_onestep, x_onestep_target, atol=tol)
-
-            x_twosteps = bunch.adag_opt.step(f, x_onestep)
-            past_grads = (
-                gradf(x_start)[0] * gradf(x_start)[0] + gradf(x_onestep)[0] * gradf(x_onestep)[0]
-            )
-            adapt_stepsize = stepsize / np.sqrt(past_grads + 1e-8)
-            x_twosteps_target = x_onestep - gradf(x_onestep)[0] * adapt_stepsize
-            assert np.allclose(x_twosteps, x_twosteps_target, atol=tol)
-
-    def test_adagrad_optimizer_multivar(self, bunch, tol):
-        """Tests that adagrad optimizer takes one and two steps correctly
-        for multivariate functions."""
-
-        for gradf, f, name in zip(grad_multi_funcs, multivariate_funcs, fnames):
-            for jdx in range(len(x_vals[:-1])):
-                bunch.adag_opt.reset()
-
-                x_vec = x_vals[jdx : jdx + 2]
-                x_onestep = bunch.adag_opt.step(f, x_vec)
-                past_grads = gradf(x_vec)[0] * gradf(x_vec)[0]
-                adapt_stepsize = stepsize / np.sqrt(past_grads + 1e-8)
-                x_onestep_target = x_vec - gradf(x_vec)[0] * adapt_stepsize
-                assert np.allclose(x_onestep, x_onestep_target, atol=tol)
-
-                x_twosteps = bunch.adag_opt.step(f, x_onestep)
-                past_grads = (
-                    gradf(x_vec)[0] * gradf(x_vec)[0] + gradf(x_onestep)[0] * gradf(x_onestep)[0]
-                )
-                adapt_stepsize = stepsize / np.sqrt(past_grads + 1e-8)
-                x_twosteps_target = x_onestep - gradf(x_onestep)[0] * adapt_stepsize
-                assert np.allclose(x_twosteps, x_twosteps_target, atol=tol)
-
-    @pytest.mark.parametrize("x_start", x_vals)
-    def test_rmsprop_optimizer_univar(self, x_start, bunch, tol):
-        """Tests that rmsprop optimizer takes one and two steps correctly
-        for univariate functions."""
-
-        for gradf, f, name in zip(grad_uni_fns, univariate_funcs, fnames):
-            bunch.rms_opt.reset()
-
-            x_onestep = bunch.rms_opt.step(f, x_start)
-            past_grads = (1 - gamma) * gradf(x_start)[0] * gradf(x_start)[0]
-            adapt_stepsize = stepsize / np.sqrt(past_grads + 1e-8)
-            x_onestep_target = x_start - gradf(x_start)[0] * adapt_stepsize
-            assert np.allclose(x_onestep, x_onestep_target, atol=tol)
-
-            x_twosteps = bunch.rms_opt.step(f, x_onestep)
-            past_grads = (1 - gamma) * gamma * gradf(x_start)[0] * gradf(x_start)[0] + (
-                1 - gamma
-            ) * gradf(x_onestep)[0] * gradf(x_onestep)[0]
-            adapt_stepsize = stepsize / np.sqrt(past_grads + 1e-8)
-            x_twosteps_target = x_onestep - gradf(x_onestep)[0] * adapt_stepsize
-            assert np.allclose(x_twosteps, x_twosteps_target, atol=tol)
-
-    def test_rmsprop_optimizer_multivar(self, bunch, tol):
-        """Tests that rmsprop optimizer takes one and two steps correctly
-        for multivariate functions."""
-
-        for gradf, f, name in zip(grad_multi_funcs, multivariate_funcs, fnames):
-            for jdx in range(len(x_vals[:-1])):
-                bunch.rms_opt.reset()
-
-                x_vec = x_vals[jdx : jdx + 2]
-                x_onestep = bunch.rms_opt.step(f, x_vec)
-                past_grads = (1 - gamma) * gradf(x_vec)[0] * gradf(x_vec)[0]
-                adapt_stepsize = stepsize / np.sqrt(past_grads + 1e-8)
-                x_onestep_target = x_vec - gradf(x_vec)[0] * adapt_stepsize
-                assert np.allclose(x_onestep, x_onestep_target, atol=tol)
-
-                x_twosteps = bunch.rms_opt.step(f, x_onestep)
-                past_grads = (1 - gamma) * gamma * gradf(x_vec)[0] * gradf(x_vec)[0] + (
-                    1 - gamma
-                ) * gradf(x_onestep)[0] * gradf(x_onestep)[0]
-                adapt_stepsize = stepsize / np.sqrt(past_grads + 1e-8)
-                x_twosteps_target = x_onestep - gradf(x_onestep)[0] * adapt_stepsize
-                assert np.allclose(x_twosteps, x_twosteps_target, atol=tol)
-
-    @pytest.mark.parametrize("x_start", x_vals)
-    def test_adam_optimizer_univar(self, x_start, bunch, tol):
-        """Tests that adam optimizer takes one and two steps correctly
-        for univariate functions."""
-
-        for gradf, f, name in zip(grad_uni_fns, univariate_funcs, fnames):
-            bunch.adam_opt.reset()
-
-            x_onestep = bunch.adam_opt.step(f, x_start)
-            adapted_stepsize = stepsize * np.sqrt(1 - delta) / (1 - gamma)
-            firstmoment = gradf(x_start)[0]
-            secondmoment = gradf(x_start)[0] * gradf(x_start)[0]
-            x_onestep_target = x_start - adapted_stepsize * firstmoment / (
-                np.sqrt(secondmoment) + 1e-8
-            )
-            assert np.allclose(x_onestep, x_onestep_target, atol=tol)
-
-            x_twosteps = bunch.adam_opt.step(f, x_onestep)
-            adapted_stepsize = stepsize * np.sqrt(1 - delta ** 2) / (1 - gamma ** 2)
-            firstmoment = gamma * gradf(x_start)[0] + (1 - gamma) * gradf(x_onestep)[0]
-            secondmoment = (
-                delta * gradf(x_start)[0] * gradf(x_start)[0]
-                + (1 - delta) * gradf(x_onestep)[0] * gradf(x_onestep)[0]
-            )
-            x_twosteps_target = x_onestep - adapted_stepsize * firstmoment / (
-                np.sqrt(secondmoment) + 1e-8
-            )
-            assert np.allclose(x_twosteps, x_twosteps_target, atol=tol)
-
-    def test_adam_optimizer_multivar(self, bunch, tol):
-        """Tests that adam optimizer takes one and two steps correctly
-        for multivariate functions."""
-
-        for gradf, f, name in zip(grad_multi_funcs, multivariate_funcs, fnames):
-            for jdx in range(len(x_vals[:-1])):
-                bunch.adam_opt.reset()
-
-                x_vec = x_vals[jdx : jdx + 2]
-                x_onestep = bunch.adam_opt.step(f, x_vec)
-                adapted_stepsize = stepsize * np.sqrt(1 - delta) / (1 - gamma)
-                firstmoment = gradf(x_vec)[0]
-                secondmoment = gradf(x_vec)[0] * gradf(x_vec)[0]
-                x_onestep_target = x_vec - adapted_stepsize * firstmoment / (
-                    np.sqrt(secondmoment) + 1e-8
-                )
-                assert np.allclose(x_onestep, x_onestep_target, atol=tol)
-
-                x_twosteps = bunch.adam_opt.step(f, x_onestep)
-                adapted_stepsize = stepsize * np.sqrt(1 - delta ** 2) / (1 - gamma ** 2)
-                firstmoment = gamma * gradf(x_vec)[0] + (1 - gamma) * gradf(x_onestep)[0]
-                secondmoment = (
-                    delta * gradf(x_vec)[0] * gradf(x_vec)[0]
-                    + (1 - delta) * gradf(x_onestep)[0] * gradf(x_onestep)[0]
-                )
-                x_twosteps_target = x_onestep - adapted_stepsize * firstmoment / (
-                    np.sqrt(secondmoment) + 1e-8
-                )
-                assert np.allclose(x_twosteps, x_twosteps_target, atol=tol)
-
-    def test_adam_optimizer_properties(self, bunch):
-        """Test the adam property interfaces"""
-        bunch.adam_opt.reset()
-        # check if None is returned when accumulation is empty
-        assert bunch.adam_opt.fm == None
-        assert bunch.adam_opt.sm == None
-        assert bunch.adam_opt.t == None
-
-        # Do some calculations to fill accumulation
-        bunch.adam_opt.step(np.sin, np.random.rand(1))
-
-        # Check the properties return the same values, stored in accumulation
-        assert bunch.adam_opt.fm == bunch.adam_opt.accumulation["fm"]
-        assert bunch.adam_opt.sm == bunch.adam_opt.accumulation["sm"]
-        assert bunch.adam_opt.t == bunch.adam_opt.accumulation["t"]
-
-    @staticmethod
-    def rotosolve_step(f, x):
-        """Helper function to test the Rotosolve and Rotoselect optimizers"""
-        # make sure that x is an array
-        if np.ndim(x) == 0:
-            x = np.array([x])
-
-        # helper function for x[d] = theta
-        def insert(xf, d, theta):
-            xf[d] = theta
-            return xf
-
-        for d, _ in enumerate(x):
-            H_0 = float(f(insert(x, d, 0)))
-            H_p = float(f(insert(x, d, np.pi / 2)))
-            H_m = float(f(insert(x, d, -np.pi / 2)))
-            a = onp.arctan2(2 * H_0 - H_p - H_m, H_p - H_m)
-
-            x[d] = -np.pi / 2 - a
-
-            if x[d] <= -np.pi:
-                x[d] += 2 * np.pi
-        return x
-
-    @pytest.mark.parametrize("x_start", [[1.2, 0.2], [-0.62, -2.1], [0.05, 0.8]])
-    @pytest.mark.parametrize(
-        "generators", [list(tup) for tup in it.product([qml.RX, qml.RY, qml.RZ], repeat=2)]
-    )
-    def test_rotoselect_optimizer(self, x_start, generators, bunch, tol):
-        """Tests that rotoselect optimizer finds the optimal generators and parameters for the
-        VQE circuit defined in `this rotoselect tutorial
-        <https://pennylane.ai/qml/demos/tutorial_rotoselect.html>`_."""
-
-        # the optimal generators for the 2-qubit VQE circuit
-        # H = 0.5 * Y_2 + 0.8 * Z_1 - 0.2 * X_1
-        optimal_generators = [qml.RY, qml.RX]
-        possible_generators = [qml.RX, qml.RY, qml.RZ]
-        bunch.rotoselect_opt.possible_generators = possible_generators
-
-        dev = qml.device("default.qubit", shots=None, wires=2)
-
-        def ansatz(params, generators):
-            generators[0](params[0], wires=0)
-            generators[1](params[1], wires=1)
-            qml.CNOT(wires=[0, 1])
-
-        @qml.qnode(dev)
-        def circuit_1(params, generators=None):  # generators will be passed as a keyword arg
-            ansatz(params, generators)
-            return qml.expval(qml.PauliZ(0)), qml.expval(qml.PauliY(1))
-
-        @qml.qnode(dev)
-        def circuit_2(params, generators=None):  # generators will be passed as a keyword arg
-            ansatz(params, generators)
-            return qml.expval(qml.PauliX(0))
-
-        def cost_fn(params, generators):
-            Z_1, Y_2 = circuit_1(params, generators=generators)
-            X_1 = circuit_2(params, generators=generators)
-            return 0.5 * Y_2 + 0.8 * Z_1 - 0.2 * X_1
-
-        f_best_gen = lambda x: cost_fn(x, optimal_generators)
-        optimal_x_start = x_start.copy()
-
-        # after four steps the optimzer should find the optimal generators/x_start values
-        for _ in range(4):
-            x_start, generators = bunch.rotoselect_opt.step(cost_fn, x_start, generators)
-            optimal_x_start = self.rotosolve_step(f_best_gen, optimal_x_start)
-
-        assert np.allclose(x_start, optimal_x_start, atol=tol)
-        assert generators == optimal_generators
-
-    @pytest.mark.parametrize("x_start", [[1.2, 0.2], [-0.62, -2.1], [0.05, 0.8]])
-    def test_keywords_rotoselect(self, bunch, x_start, tol):
-        """test rotoselect accepts keywords"""
-
-        generators = [qml.RY, qml.RX]
-        possible_generators = [qml.RX, qml.RY, qml.RZ]
-        bunch.rotoselect_opt.possible_generators = possible_generators
-
-        dev = qml.device("default.qubit", shots=None, wires=2)
-
-        def ansatz(params, generators):
-            generators[0](params[0], wires=0)
-            generators[1](params[1], wires=1)
-            qml.CNOT(wires=[0, 1])
-
-        @qml.qnode(dev)
-        def circuit_1(params, generators=None):  # generators will be passed as a keyword arg
-            ansatz(params, generators)
-            return qml.expval(qml.PauliZ(0)), qml.expval(qml.PauliY(1))
-
-        @qml.qnode(dev)
-        def circuit_2(params, generators=None):  # generators will be passed as a keyword arg
-            ansatz(params, generators)
-            return qml.expval(qml.PauliX(0))
-
-        def cost_fn(params, generators, shift=0.0):
-            Z_1, Y_2 = circuit_1(params, generators=generators)
-            X_1 = circuit_2(params, generators=generators)
-            return 0.5 * (Y_2 - shift) ** 2 + 0.8 * (Z_1 - shift) ** 2 - 0.2 * (X_1 - shift) ** 2
-
-        params_new, _, res_new = bunch.rotoselect_opt.step_and_cost(
-            cost_fn, x_start, generators, shift=0.0
-        )
-        params_new2, _, res_new2 = bunch.rotoselect_opt.step_and_cost(
-            cost_fn, x_start, generators, shift=1.0
-        )
-
-        assert not np.allclose(params_new, params_new2, atol=tol)
-        assert np.allclose(res_new2, cost_fn(x_start, generators, shift=1.0), atol=tol)
-
-    def test_update_stepsize(self):
-        """
-        Tests whether the stepsize value is updated correctly and whether a ``UserWarning``
-        is raised when the ``update_stepsize`` method is used.
-        """
-
-        eta = 0.5
-        opt = AdamOptimizer(eta)
-        assert opt.stepsize == eta
-
-        eta2 = 0.1
-        opt.update_stepsize(eta2)
-        assert opt.stepsize == eta2
-
-        with pytest.warns(
-            UserWarning,
-            match="'update_stepsize' is deprecated. Stepsize value can be updated using "
-            "the 'stepsize' attribute.",
-        ):
-            opt.update_stepsize(eta)
-
-    def test_private_stepsize(self):
-        """
-        Tests whether it is possible to get and set ``stepsize`` using ``_stepsize``
-        and whether a ``UserWarning`` is raised while doing so.
-        """
-        eta = 0.5
-        opt = AdamOptimizer(eta)
-        assert opt._stepsize == eta
-
-        with pytest.warns(
-            UserWarning, match="'_stepsize' is deprecated. Please use 'stepsize' instead."
-        ):
-            opt._stepsize
-
-        eta2 = 0.1
-        opt._stepsize = eta2
-        assert opt.stepsize == eta2
-
-        with pytest.warns(
-            UserWarning, match="'_stepsize' is deprecated. Please use 'stepsize' instead."
-        ):
-            opt._stepsize = eta2
-
-
-def reset(opt):
-    if getattr(opt, "reset", None):
-        opt.reset()
 
 
 @pytest.fixture
@@ -737,12 +77,12 @@ class TestOverOpts:
         spy = mocker.spy(wrapper, "func")
 
         x_new_two = opt.step(wrapper.func, x, c=2.0)
-        reset(opt)
+        self.reset(opt)
 
         args2, kwargs2 = spy.call_args_list[-1]
 
         x_new_three_wc, cost_three = opt.step_and_cost(wrapper.func, x, c=3.0)
-        reset(opt)
+        self.reset(opt)
 
         args3, kwargs3 = spy.call_args_list[-1]
 
@@ -774,11 +114,11 @@ class TestOverOpts:
         z = np.array([3.0])
 
         (x_new, y_new, z_new), cost = opt.step_and_cost(wrapper.func, x, y, z)
-        reset(opt)
+        self.reset(opt)
         args_called1, kwargs1 = spy.call_args_list[-1]  # just take last call
 
         x_new2, y_new2, z_new2 = opt.step(wrapper.func, x_new, y_new, z_new)
-        reset(opt)
+        self.reset(opt)
         args_called2, kwargs2 = spy.call_args_list[-1]  # just take last call
 
         if opt_name != "roto":
@@ -804,9 +144,9 @@ class TestOverOpts:
         data = np.array([1.0], requires_grad=False)
 
         args_new = opt.step(func, x, data)
-        reset(opt)
+        self.reset(opt)
         args_new_wc, cost = opt.step_and_cost(func, *args_new)
-        reset(opt)
+        self.reset(opt)
 
         assert np.allclose(len(args_new), 2, atol=tol)
         assert not np.allclose(args_new[0], x, atol=tol)
@@ -830,10 +170,10 @@ class TestOverOpts:
         args = (x, y, z)
 
         x_seperate, y_seperate, z_seperate = opt.step(func1, x, y, z)
-        reset(opt)
+        self.reset(opt)
 
         args_new = opt.step(func2, args)
-        reset(opt)
+        self.reset(opt)
 
         assert np.allclose(x_seperate, args_new[0], atol=tol)
         assert np.allclose(y_seperate, args_new[1], atol=tol)
@@ -911,3 +251,8 @@ class TestOverOpts:
         # check that the argument to RX doesn't change, as the X rotation doesn't influence <Z>
         assert x == 0
         assert ev == original_ev
+
+    @staticmethod
+    def reset(opt):
+        if getattr(opt, "reset", None):
+            opt.reset()
